train ticket and theater ticket classes
